+++ openwisp_controller/urls.py
@@ -32,6 +32,12 @@ url_metadata = [
         'app': 'openwisp_controller.geo',
         'include': {'module': '{app}.api.urls', 'namespace': 'geo'},
     },
+    # openwisp_controller.config
+    {
+        'regexp': r'^',
+        'app': 'openwisp_controller.config',
+        'include': {'module': '{app}.urls', 'namespace': 'config'},
+    },
 ]
 
 urlpatterns = []